Rename webgl-new to webgl
